requirements.txt
Flask==3.1.0
Pillow==9.0.0
Werkzeug==3.1.3
